+++ libgo/go/runtime/heapdump.go
@@ -364,8 +364,8 @@ func dumpobjs() {
 			throw("freemark array doesn't have enough entries")
 		}
 
-		for freeIndex := uint16(0); freeIndex < s.nelems; freeIndex++ {
-			if s.isFree(uintptr(freeIndex)) {
+		for freeIndex := uintptr(0); freeIndex < s.nelems; freeIndex++ {
+			if s.isFree(freeIndex) {
 				freemark[freeIndex] = true
 			}
 		}
